Reworked the replacing system
Now the replacing system is much more cleaner. It is also reordered to
work well with all the features. Local variables now have better support
and will be replaced in most cases (if not, please file a bug report). I
fixed an unknown bug in `libbutter.butter`. Soon I'll be working in a
busted test suite.